fix check tests
