+++ packages/cli/tests/build.rs
@@ -597,38 +597,6 @@ async fn file_template_raw() {
 }
 
 #[tokio::test]
-async fn blob_rejects_non_strings() {
-	let directory = temp::directory! {
-		"tangram.ts" => r#"
-			import file from "./file.txt";
-			export default () => tg.blob`\n\t${file}\n`;
-		"#,
-		"file.txt" => "Hello, world!"
-	};
-	let assertions = |output: std::process::Output| async move {
-		assert_failure!(output);
-	};
-	let path = "";
-	test_check(directory, path, assertions).await;
-}
-
-#[tokio::test]
-async fn file_rejects_non_strings() {
-	let directory = temp::directory! {
-		"tangram.ts" => r#"
-			import file from "./file.txt";
-			export default () => tg.file`\n\t${file}\n`;
-		"#,
-		"file.txt" => "Hello, world!"
-	};
-	let assertions = |output: std::process::Output| async move {
-		assert_failure!(output);
-	};
-	let path = "";
-	test_check(directory, path, assertions).await;
-}
-
-#[tokio::test]
 async fn directory_get_follows_intermediate_component_symlinks() {
 	let directory = temp::directory! {
 		"tangram.ts" => indoc!(r#"
@@ -1152,34 +1120,6 @@ async fn test_build_remote<F, Fut>(
 		assert_success!(remote_output);
 
 		assertions(local_output, remote_output).await;
-	})
-	.await;
-}
-
-async fn test_check<F, Fut>(
-	artifact: impl Into<temp::Artifact> + Send + 'static,
-	path: &str,
-	assertions: F,
-) where
-	F: FnOnce(std::process::Output) -> Fut + Send + 'static,
-	Fut: Future<Output = ()> + Send,
-{
-	test(TG, async move |context| {
-		let server = context.spawn_server().await.unwrap();
-
-		let artifact: temp::Artifact = artifact.into();
-		let temp = Temp::new();
-		artifact.to_path(temp.as_ref()).await.unwrap();
-
-		let path = temp.path().join(path);
-		let command_ = format!("{path}", path = path.display());
-
-		// Build.
-		let mut command = server.tg();
-		command.arg("check").arg(command_);
-		let output = command.output().await.unwrap();
-
-		assertions(output).await;
 	})
 	.await;
 }
